+++ src/from.rs
@@ -1,8 +1,8 @@
 //! Create arrow-rs objects from an `Robj`
 //!
 //! ```ignore
-//! fn array_from_r(field: Robj) -> Result<ArrayData> {
-//!     ArrayData::from_arrow_robj(&field)?
+//! fn array_from_r(field: Robj) -> Result<ArrayData, anyhow::Error> {
+//!     ArrayData::from_arrow_robj(&field)
 //! }
 //! ```
 //!
@@ -23,26 +23,16 @@
 //! the first chunk is returned. If you expect more than one chunk, use `ArrowArrayStreamReader`.
 //!
 
+use crate::FromArrowRobj;
+use anyhow::Result;
 use arrow::{
     array::{make_array, ArrayData},
     datatypes::{DataType, Field, Schema},
-    error::ArrowError,
     ffi::{self, FFI_ArrowArray, FFI_ArrowSchema},
     ffi_stream::{self, ArrowArrayStreamReader, FFI_ArrowArrayStream},
     record_batch::RecordBatch,
 };
-
 use extendr_api::prelude::*;
-use std::result::Result;
-
-/// Creates arrow-rs Structs from an Robj
-///
-
-pub trait FromArrowRobj: Sized {
-    fn from_arrow_robj(robj: &Robj) -> Result<Self, ErrArrowRobj>;
-}
-
-pub type ErrArrowRobj = ArrowError;
 
 /// Calls `nanoarrow::nanoarrow_pointer_addr_chr()`
 ///
@@ -69,24 +59,17 @@ pub fn nanoarrow_export(source: &Robj, dest: String) -> Result<Robj, Error> {
 }
 
 impl FromArrowRobj for Field {
-    fn from_arrow_robj(robj: &Robj) -> Result<Self, ErrArrowRobj> {
-        // handle nanoarrow
+    fn from_arrow_robj(robj: &Robj) -> std::result::Result<Self, anyhow::Error> {
         if robj.inherits("nanoarrow_schema") {
             let c_schema = FFI_ArrowSchema::empty();
             let c_schema_ptr = &c_schema as *const FFI_ArrowSchema as usize;
-
             let _ = nanoarrow_export(robj, c_schema_ptr.to_string());
-
-            let field = Field::try_from(&c_schema)?;
-
-            return Ok(field);
+            return Ok(Field::try_from(&c_schema)?);
         }
 
-        let is_field = robj.inherits("Field");
-
-        if !(is_field) {
-            return Err(ErrArrowRobj::ParseError(
-                "did not find a `Field` or `nanoarrow_schema`".into(),
+        if !robj.inherits("Field") {
+            return Err(anyhow::anyhow!(
+                "did not find a `Field` or `nanoarrow_schema`"
             ));
         }
 
@@ -98,32 +81,23 @@ impl FromArrowRobj for Field {
 
         let c_schema = FFI_ArrowSchema::empty();
         let c_schema_ptr = &c_schema as *const FFI_ArrowSchema as usize;
-
         let _ = export_to_c.call(pairlist!(c_schema_ptr.to_string()));
-        let field = Field::try_from(&c_schema)?;
-
-        Ok(field)
+        Ok(Field::try_from(&c_schema)?)
     }
 }
 
 impl FromArrowRobj for DataType {
-    fn from_arrow_robj(robj: &Robj) -> Result<Self, ErrArrowRobj> {
+    fn from_arrow_robj(robj: &Robj) -> std::result::Result<Self, anyhow::Error> {
         if robj.inherits("nanoarrow_schema") {
             let c_schema = FFI_ArrowSchema::empty();
             let c_schema_ptr = &c_schema as *const FFI_ArrowSchema as usize;
-
             let _ = nanoarrow_export(robj, c_schema_ptr.to_string());
-
-            let field = DataType::try_from(&c_schema)?;
-
-            return Ok(field);
+            return Ok(DataType::try_from(&c_schema)?);
         }
 
-        let is_datatype = robj.inherits("DataType");
-
-        if !(is_datatype) {
-            return Err(ErrArrowRobj::ParseError(
-                "did not find a `DataType` or `nanoarrow_schema`".into(),
+        if !robj.inherits("DataType") {
+            return Err(anyhow::anyhow!(
+                "did not find a `DataType` or `nanoarrow_schema`"
             ));
         }
 
@@ -135,32 +109,23 @@ impl FromArrowRobj for DataType {
 
         let c_schema = FFI_ArrowSchema::empty();
         let c_schema_ptr = &c_schema as *const FFI_ArrowSchema as usize;
-
         let _ = export_to_c.call(pairlist!(c_schema_ptr.to_string()));
-        let data_type = DataType::try_from(&c_schema)?;
-
-        Ok(data_type)
+        Ok(DataType::try_from(&c_schema)?)
     }
 }
 
 impl FromArrowRobj for Schema {
-    fn from_arrow_robj(robj: &Robj) -> Result<Self, ErrArrowRobj> {
+    fn from_arrow_robj(robj: &Robj) -> std::result::Result<Self, anyhow::Error> {
         if robj.inherits("nanoarrow_schema") {
             let c_schema = FFI_ArrowSchema::empty();
             let c_schema_ptr = &c_schema as *const FFI_ArrowSchema as usize;
-
             let _ = nanoarrow_export(robj, c_schema_ptr.to_string());
-
-            let field = Schema::try_from(&c_schema)?;
-
-            return Ok(field);
+            return Ok(Schema::try_from(&c_schema)?);
         }
 
-        let is_schema = robj.inherits("Schema");
-
-        if !(is_schema) {
-            return Err(ErrArrowRobj::ParseError(
-                "did not find a `Schema` or `nanoarrow_schema`".into(),
+        if !robj.inherits("Schema") {
+            return Err(anyhow::anyhow!(
+                "did not find a `Schema` or `nanoarrow_schema`"
             ));
         }
 
@@ -172,17 +137,14 @@ impl FromArrowRobj for Schema {
 
         let c_schema = FFI_ArrowSchema::empty();
         let c_schema_ptr = &c_schema as *const FFI_ArrowSchema as usize;
-
         let _ = export_to_c.call(pairlist!(c_schema_ptr.to_string()));
-        let schema = Schema::try_from(&c_schema)?;
-
-        Ok(schema)
+        Ok(Schema::try_from(&c_schema)?)
     }
 }
 
 // https://github.com/apache/arrow-rs/blob/200e8c80084442d9579e00967e407cd83191565d/arrow/src/pyarrow.rs#L248
 impl FromArrowRobj for ArrayData {
-    fn from_arrow_robj(robj: &Robj) -> Result<Self, ErrArrowRobj> {
+    fn from_arrow_robj(robj: &Robj) -> std::result::Result<Self, anyhow::Error> {
         if robj.inherits("nanoarrow_array") {
             let array = FFI_ArrowArray::empty();
             let schema = FFI_ArrowSchema::empty();
@@ -200,16 +162,13 @@ impl FromArrowRobj for ArrayData {
             let _ = nanoarrow_export(robj, c_array_ptr.to_string());
             let _ = nanoarrow_export(&robj_schema, c_schema_ptr.to_string());
 
-            return unsafe { ffi::from_ffi(array, &schema) };
+            return Ok(unsafe { ffi::from_ffi(array, &schema)? });
         }
 
-        let is_array = robj.inherits("Array");
-
-        if !is_array {
-            return Err(ErrArrowRobj::ParseError("did not find a `Array`".into()));
+        if !robj.inherits("Array") {
+            return Err(anyhow::anyhow!("did not find a `Array`"));
         }
 
-        // prepare a pointer to receive the Array struct
         let array = FFI_ArrowArray::empty();
         let schema = FFI_ArrowSchema::empty();
 
@@ -224,36 +183,29 @@ impl FromArrowRobj for ArrayData {
 
         let _ = export_to_c.call(pairlist!(c_array_ptr.to_string(), c_schema_ptr.to_string()));
 
-        unsafe { ffi::from_ffi(array, &schema) }
+        Ok(unsafe { ffi::from_ffi(array, &schema)? })
     }
 }
 
-/// If there are more than one RecordBatches in the stream, do not use this
-/// Use ArrowStreamReader instead
+/// If there are more than one RecordBatches in the stream, do not use this.
+/// Use `ArrowArrayStreamReader` instead.
 impl FromArrowRobj for RecordBatch {
-    fn from_arrow_robj(robj: &Robj) -> Result<Self, ErrArrowRobj> {
+    fn from_arrow_robj(robj: &Robj) -> std::result::Result<Self, anyhow::Error> {
         if robj.inherits("nanoarrow_array_stream") {
-            // we need to allocate an empty schema and fetch it from the record batch
             let stream = ffi_stream::FFI_ArrowArrayStream::empty();
             let c_stream_ptr = &stream as *const FFI_ArrowArrayStream as usize;
-
             let _ = nanoarrow_export(robj, c_stream_ptr.to_string());
-
             let res = ArrowArrayStreamReader::try_new(stream)?;
             let r2 = res.into_iter().map(|xi| xi.unwrap()).nth(0).unwrap();
-
             return Ok(r2);
         }
 
-        let is_rb = robj.inherits("RecordBatch");
-
-        if !is_rb {
-            return Err(ErrArrowRobj::ParseError(
-                "did not find a `RecordBatch` or `nanoarrow_array_stream`".into(),
+        if !robj.inherits("RecordBatch") {
+            return Err(anyhow::anyhow!(
+                "did not find a `RecordBatch` or `nanoarrow_array_stream`"
             ));
         }
 
-        // we need to allocate an empty schema and fetch it from the record batch
         let array = FFI_ArrowArray::empty();
         let schema = FFI_ArrowSchema::empty();
 
@@ -273,30 +225,22 @@ impl FromArrowRobj for RecordBatch {
 
         let res_arrays = res
             .child_data()
-            .into_iter()
+            .iter()
             .map(|xi| make_array(xi.clone()))
             .collect::<Vec<_>>();
 
-        let res = RecordBatch::try_new(schema.into(), res_arrays)?;
-
-        Ok(res)
+        Ok(RecordBatch::try_new(schema.into(), res_arrays)?)
     }
 }
 
 impl FromArrowRobj for ArrowArrayStreamReader {
-    fn from_arrow_robj(robj: &Robj) -> Result<Self, ErrArrowRobj> {
-        // TODO arrow::RecordBatchStreamWriter
+    fn from_arrow_robj(robj: &Robj) -> std::result::Result<Self, anyhow::Error> {
         if !robj.inherits("nanoarrow_array_stream") {
-            return Err(ErrArrowRobj::ParseError(
-                "did not find `nanoarrow_array_stream`".into(),
-            ));
+            return Err(anyhow::anyhow!("did not find `nanoarrow_array_stream`"));
         }
-        // we need to allocate an empty schema and fetch it from the record batch
         let stream = ffi_stream::FFI_ArrowArrayStream::empty();
         let c_stream_ptr = &stream as *const FFI_ArrowArrayStream as usize;
-
         let _ = nanoarrow_export(robj, c_stream_ptr.to_string());
-
-        ArrowArrayStreamReader::try_new(stream)
+        Ok(ArrowArrayStreamReader::try_new(stream)?)
     }
 }